refactor: move 3rd party plugin service
