+++ packages/mask/shared/definitions/routes.ts
@@ -9,3 +9,15 @@ export function getPermissionRequestURL(permission: browser.permissions.Permissi
     for (const each of permissions) params.append('permissions', each)
     return `${getPopupRouteURLWithNoParam(PopupRoutes.RequestPermission)}?${params.toString()}`
 }
+export function constructThirdPartyRequestPermissionURL(
+    pluginManifestURL: string,
+    permissions: ThirdPartyPluginPermission[],
+) {
+    const params = new URLSearchParams()
+    params.set('plugin', pluginManifestURL)
+    for (const x of permissions) params.append('permission', String(x))
+    return getPopupRouteURLWithNoParam(PopupRoutes.ThirdPartyRequestPermission) + '?' + params.toString()
+}
+export enum ThirdPartyPluginPermission {
+    DEBUG_Profiles = 1,
+}
