--- packages/mask/src/extension/background-script/ThirdPartyPlugin/types.ts
@@ -1,3 +1,0 @@
-export enum ThirdPartyPluginPermission {
-    DEBUG_Profiles = 1,
-}
